backend/requirements.txt
torch
transformers
uvicorn
fastapi
sentencepiece
pymysql
starlette-session
cryptography
websockets
pytest